Move args_handler and path_handler to CLI module
